+++ bin/compare.mli
@@ -1,4 +1,3 @@
 open! Core
-open Core_extended.Std
 
-val command : Deprecated_command.t
+val command : Command.t